Refactoring: Extract selection dialog to separate class
